Changes how login/create work
