+++ frontend/app/package.json
@@ -104,6 +104,7 @@
     "vite-plugin-checker": "0.5.1",
     "vite-plugin-istanbul": "3.0.2",
     "vite-plugin-pwa": "0.13.3",
+    "vite-plugin-vue-layouts": "0.7.0",
     "vitest": "0.25.2",
     "vue-eslint-parser": "9.1.0",
     "vue-i18n-extract": "2.0.7",
