+++ frontend/app/src/components/user/LoginHeader.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="pa-6 text-h2 font-weight-black white--text primary">
+    <div>{{ t('app.name') }}</div>
+    <div class="text-caption">
+      {{ t('app.moto') }}
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+const { t } = useI18n();
+</script>
