+++ frontend/app/src/composables/premium.ts
@@ -1,6 +1,43 @@
+import { useInterop } from '@/electron-interop';
+import { useMessageStore } from '@/store/message';
+import { useSessionAuthStore } from '@/store/session/auth';
 import { usePremiumStore } from '@/store/session/premium';
 
 export const usePremium = () => {
   const { premium } = storeToRefs(usePremiumStore());
   return premium;
+};
+
+export const usePremiumReminder = () => {
+  const premium = usePremium();
+  const { premiumUserLoggedIn } = useInterop();
+  const { premiumPrompt } = storeToRefs(useSessionAuthStore());
+  const { message } = storeToRefs(useMessageStore());
+
+  const isPremiumDialogVisible = computed(
+    () => !get(premium) && !get(message).title && get(premiumPrompt)
+  );
+
+  const showPremiumDialog = () => {
+    if (get(premium)) {
+      set(premiumPrompt, false);
+      return;
+    }
+    set(premiumPrompt, true);
+  };
+
+  const dismissPremiumDialog = () => {
+    set(premiumPrompt, false);
+  };
+
+  const showGetPremiumButton = () => {
+    premiumUserLoggedIn(get(premium));
+  };
+
+  return {
+    isPremiumDialogVisible,
+    showGetPremiumButton,
+    showPremiumDialog,
+    dismissPremiumDialog
+  };
 };
