+++ frontend/app/src/store/session/auth.ts
@@ -1,0 +1,33 @@
+import { SyncConflict } from '@/store/session/types';
+
+const defaultSyncConflict = (): SyncConflict => ({
+  message: '',
+  payload: null
+});
+
+export const useSessionAuthStore = defineStore('session/auth', () => {
+  const newAccount = ref(false);
+  const logged = ref(false);
+  const loginComplete = ref(false);
+  const premiumPrompt = ref(false);
+  const username = ref('');
+  const syncConflict = ref<SyncConflict>(defaultSyncConflict());
+
+  const resetSyncConflict = () => {
+    set(syncConflict, defaultSyncConflict());
+  };
+
+  return {
+    newAccount,
+    logged,
+    loginComplete,
+    username,
+    premiumPrompt,
+    syncConflict,
+    resetSyncConflict
+  };
+});
+
+if (import.meta.hot) {
+  import.meta.hot.accept(acceptHMRUpdate(useSessionAuthStore, import.meta.hot));
+}
